+++ NEWS.md
@@ -3,3 +3,4 @@
 * [KAFKAWRAP-10](https://issues.folio.org/browse/KAFKAWRAP-10) Provide property to set compression type for producer configuration
 * [KAFKAWRAP-11](https://issues.folio.org/browse/KAFKAWRAP-11) Remove duplicate kafka headers
 * [MODDATAIMP-557](https://issues.folio.org/browse/MODDATAIMP-557) Close Vertx Kafka producers that have not been closed
+* [MODDATAIMP-623](https://issues.folio.org/browse/MODDATAIMP-623) Remove Kafka cache initialization and Maven dependency